+++ src/components/DepositWithdraw/ApproveToken/index.scss
@@ -1,4 +1,4 @@
-@import "../constants.scss";
+@import "../../../styles/constants.scss";
 
 .approve-token-backdrop {
   position: fixed;
