+++ src/components/DepositWithdraw/Assets/index.scss
@@ -11,6 +11,7 @@
   }
   th, td {
     padding: 15px 0px;
+    background-color: unset;
   }
   .assets-column {
     width: 70%;
