+++ src/components/DepositWithdraw/PendingWithdraw/index.scss
@@ -1,0 +1,13 @@
+@import "../../../styles/constants.scss";
+
+.pending-withdraw-timer-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: $white;
+  margin-bottom: 50px;
+  > span {
+    margin: 30px 0;
+    text-transform: uppercase;
+  }
+}
