+++ src/components/DepositWithdraw/index.scss
@@ -1,4 +1,4 @@
-@import "../constants.scss";
+@import "../../styles/constants.scss";
 @import "rc-tooltip/assets/bootstrap_white.css";
 
 .deposit-wrapper {
@@ -39,17 +39,6 @@
         border-bottom: 2px solid $turquoise;
         font-weight: bold;
       }
-    }
-  }
-  .pending-withdraw-timer-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: $white;
-    margin-bottom: 50px;
-    > span {
-      margin: 30px 0;
-      text-transform: uppercase;
     }
   }
   .balance-wrapper {
@@ -151,13 +140,6 @@
       display: flex;
       align-items: center;
     }
-  }
-  .withdraw-timer-wrapper {
-    display: flex;
-    justify-content: space-between;
-    color: $white;
-    padding-top: 20px;
-    padding-bottom: 20px;
   }
   .terms-of-use-wrapper {
     color: $white;
